separated auth and user controllers
