+++ views/user/show.ejs
@@ -1,0 +1,137 @@
+<%- include("../partials/header.ejs") %>
+  <!-- Navigation -->
+  <div class="d-flex flex-column flex-md-row align-items-center p-3 px-md-4 mb-3 bg-dark border-bottom shadow-sm">
+    <h5 class="my-0 mr-md-auto font-weight-normal"><a href="/"><img class="img-fluid logo" src="/images/logo-horizontal.png" alt=""></a></h5>
+    <nav class="my-2 my-md-0 mr-md-3">
+      <a class="p-2 text-white" href="/">Menu</a>
+      <a class="p-2 text-white" href="#">Tracker</a>
+      <% if(user) { %>
+        <!-- Example single danger button -->
+      <div class="btn-group">
+        <div class="button-group dropleft">
+
+        <button class="btn btn-danger btn-sm dropdown-toggle" type="button" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
+          My Account
+        </button>
+
+        <div class="dropdown-menu">
+          <a class="dropdown-item" href="/auth/<%= user.id %>"><i class="fas fa-user"></i> View Account</a>
+          <a class="dropdown-item" href="/auth/<%= user.id %>/edit"><i class="fas fa-user-edit"></i> Edit Account</a>
+          <a class="dropdown-item" href="/orders/"><i class="fas fa-pizza-slice"></i> View Orders</a>
+          <div class="dropdown-divider"></div>
+          <div class="dropdown-item">
+            <form action="/logout?_method=DELETE" method="POST"> 
+            <button class="dropdown-item" type="submit" value="Logout" ><i class="fas fa-sign-out-alt"></i> Logout</button>
+          </form>
+          </div>
+
+        </div>
+      </div>
+      </div>
+    </nav>
+  
+    <% } else { %>
+      <div class="btn-group btn-group-toggle" data-toggle="buttons">
+      <a class="btn btn-sm btn-outline-danger" href="/login">Login</a> 
+      
+      <a class="btn btn-sm btn-outline-danger" href="/register">Register</a>
+      </div>
+      <% } %>
+
+  </div>
+  <!-- END Navigation -->
+
+<!-- Main -->
+<div class="main">
+  <div class="container">
+    <div class="row">
+      <h1>Account Info</h1>
+
+      <!-- Registration Form -->
+      <form action="/auth/<%= user._id %>?_method=PUT" method="POST">
+        <div class="container">
+          <div class="row">
+            <!-- Left Column -->
+            <div class="col">
+            <div class="container">          
+              <div class="row mt-3">
+      
+                  <div class="form-group">
+                    <label for="Address">Address</label>
+                    <input type="text" class="form-control" name="address" placeholder="Address" value="<%= user.address %> " disabled />
+                  </div>
+      
+                  <div class="col-5 mt-3">
+                    <label for="City">City</label>
+                    <input type="text" class="form-control" name="city" placeholder="City" value="<%= user.city %>" disabled />
+                  </div>
+        
+                  <div class="col-4 mt-3">
+                    <label for="State">State</label>
+                    <input type="text" class="form-control" name="state" placeholder="State" value="<%= user.state %>" disabled />
+                  </div>
+        
+                  <div class="col mt-3">
+                    <label for="Zip">Zip</label>
+                    <input type="text" class="form-control" name="zip" placeholder="Zip" value="<%= user.zip %>" disabled />
+                  </div> 
+                </div>
+              </div>
+      
+              <div class="container">
+                <div class="row mt-3">
+      
+                  <div class="col-6">
+                    <label for="cardNumber">Card Number</label>
+                    <input type="text" class="form-control" name="ccNum" placeholder="Card Number" value="<%= user.ccNum %>" disabled />
+                  </div>
+          
+                  <div class="col">
+                    <label for="CVV">CVV</label>
+                    <input type="text" class="form-control" name="ccv" placeholder="CCV" value="<%= user.ccv %>" disabled />
+                  </div>
+          
+                  <div class="col">
+                    <label for="Expirt">Expiration Date</label>
+                    <input type="text" class="form-control" name="expiry" placeholder="(MMYY)" value="<%= user.expiry %>" disabled />
+                  </div>
+
+      
+                </div>
+              </div>
+            </div>
+            <!-- END Left Column -->
+           
+            <!-- Right Column -->
+            <div class="col-4">
+              <div class="container border-left">
+                <div class="row mt-3">
+                  <div class="form-group mt-3">
+                    <label for="Username">Full Name</label>
+                    <input type="text" class="form-control" name="username" placeholder="Please enter your full name" value="<%= user.username %>" disabled />
+                  </div>
+                  <div class="form-group mt-3">
+                    <label for="Email">Email</label>
+                    <input type="email" class="form-control" name="email" placeholder="Email" value="<%= user.email %>" disabled />
+                  </div>
+                  <div class="mt-3">
+                  <a href="#">Reset Password</a>
+                </div>
+                <div class="form-group mt-3">
+                  <a name="" id="" class="btn btn-danger" href="/auth/<%= user.id %>/edit " role="button">Edit Account</a>
+                </div>
+              </div>
+            </div>
+            </div>
+            <!-- END Right Column -->
+        </div>
+        </div>
+
+      </form>
+      <!-- END Registration Form -->
+    </div>
+  </div>
+</div>
+<!-- END Main -->
+
+<%- include("../partials/footer.ejs") %>
